Move pluralize to _util.py
