+++ cloup/constraints/_core.py
@@ -7,14 +7,13 @@ from typing import (
 import click
 from click import Context, Parameter
 
-from cloup._util import check_arg, class_name, make_one_line_repr, make_repr
+from cloup._util import check_arg, class_name, make_one_line_repr, make_repr, pluralize
 from .exceptions import ConstraintViolated, UnsatisfiableConstraint
 from .common import (
     get_params_whose_value_is_set,
     get_required_params,
     join_param_labels,
     param_value_is_set,
-    pluralize,
 )
 
 Op = TypeVar('Op', bound='Operator')
